Move HashedUri over to mod signer_payload
