+++ src/identity_assertion/mod.rs
@@ -26,7 +26,7 @@ use crate::{
 };
 
 pub(crate) mod signer_payload;
-use signer_payload::SignerPayload;
+use signer_payload::{HashedUri, SignerPayload};
 
 pub(crate) mod signature_handler;
 use signature_handler::SignatureHandler;
@@ -341,38 +341,6 @@ pub enum VerifiedIdentityType {
     /// Creator Assertions Working Group and MUST NOT be used unless defined in
     /// a this or a future version of this specification.
     Other(NonEmptyString),
-}
-
-/// A `HashedUri` provides a reference to content available within the same
-/// manifest store.
-///
-/// This is described in §8.3, “[URI References],” of the C2PA Technical
-/// Specification.
-///
-/// [URI References]: https://c2pa.org/specifications/specifications/2.0/specs/C2PA_Specification.html#_uri_references
-#[derive(Clone, Deserialize, Eq, PartialEq, Serialize)]
-pub struct HashedUri {
-    /// JUMBF URI reference
-    pub url: String,
-
-    /// A string identifying the cryptographic hash algorithm used to compute
-    /// the hash
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub alg: Option<String>,
-
-    /// Byte string containing the hash value
-    #[serde(with = "serde_bytes")]
-    pub hash: Vec<u8>,
-}
-
-impl Debug for HashedUri {
-    fn fmt(&self, f: &mut Formatter<'_>) -> Result<(), std::fmt::Error> {
-        f.debug_struct("HashedUri")
-            .field("url", &self.url)
-            .field("alg", &self.alg)
-            .field("hash", &DebugByteSlice(&self.hash))
-            .finish()
-    }
 }
 
 /// Describes the ways in which a CAWG identity
