+++ src/identity_assertion/signer_payload.rs
@@ -11,11 +11,17 @@
 // specific language governing permissions and limitations under
 // each license.
 
-use std::{collections::HashSet, fmt::Debug};
+use std::{
+    collections::HashSet,
+    fmt::{Debug, Formatter},
+};
 
 use serde::{Deserialize, Serialize};
 
-use crate::identity_assertion::{HashedUri, ValidationError, ValidationResult};
+use crate::{
+    identity_assertion::{ValidationError, ValidationResult},
+    internal::debug_byte_slice::DebugByteSlice,
+};
 
 /// The set of data to be signed by the credential holder.
 #[derive(Clone, Debug, Deserialize, Eq, Serialize, PartialEq)]
@@ -91,5 +97,37 @@ impl SignerPayload {
         }
 
         Ok(())
+    }
+}
+
+/// A `HashedUri` provides a reference to content available within the same
+/// manifest store.
+///
+/// This is described in §8.3, “[URI References],” of the C2PA Technical
+/// Specification.
+///
+/// [URI References]: https://c2pa.org/specifications/specifications/2.0/specs/C2PA_Specification.html#_uri_references
+#[derive(Clone, Deserialize, Eq, PartialEq, Serialize)]
+pub struct HashedUri {
+    /// JUMBF URI reference
+    pub url: String,
+
+    /// A string identifying the cryptographic hash algorithm used to compute
+    /// the hash
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub alg: Option<String>,
+
+    /// Byte string containing the hash value
+    #[serde(with = "serde_bytes")]
+    pub hash: Vec<u8>,
+}
+
+impl Debug for HashedUri {
+    fn fmt(&self, f: &mut Formatter<'_>) -> Result<(), std::fmt::Error> {
+        f.debug_struct("HashedUri")
+            .field("url", &self.url)
+            .field("alg", &self.alg)
+            .field("hash", &DebugByteSlice(&self.hash))
+            .finish()
     }
 }
